[refactor] remove zod && use-form
